backend/pyproject.toml
[project]
name = "django-auth"
version = "0.1.0"
description = "Django authentication and authorization backend"
requires-python = ">=3.14,<3.15"
dependencies = [
    "django>=6.0.0",
    "djangorestframework>=3.0.0",
    "drf-spectacular>=0.0.0",
    "django-cors-headers>=0.0.0",
    "ldap3>=2.9",
    "python-dotenv>=0.0.0",
]

[dependency-groups]
dev = [
    "pytest>=0.0.0",
    "pytest-django>=0.0.0",
    "pytest-mock>=0.0.0",
    "pytest-cov>=0.0.0",
    "mypy>=0.0.0",
    "django-stubs>=0.0.0",
    "djangorestframework-stubs>=0.0.0",
]

[tool.uv]
default-groups = ["dev"]
